9 Ocak 2023 Authentication işlemi öncesi

+++ Entities/Concrete/Product.cs
@@ -13,7 +13,7 @@ namespace Entities.Concrete
         public int ProductId { get; set; }
         public int CategoryId { get; set; }
         public string? ProductName { get; set; }
-        public short UnitsInStock { get; set; }
+        public int UnitsInStock { get; set; }
         public string? ProductCode { get; set; }
         public string? ProductBarcode { get; set; }
         public string? ProductDescription { get; set;}
